# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CRIA - Convergent Research Intelligence Architecture API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ExperimentStatus =
  (typeof ExperimentStatus)[keyof typeof ExperimentStatus];

export const ExperimentStatus = {
  pending: "pending",
  running: "running",
  complete: "complete",
  failed: "failed",
  paused: "paused",
} as const;

export type ExperimentProject =
  (typeof ExperimentProject)[keyof typeof ExperimentProject];

export const ExperimentProject = {
  hum: "hum",
  book3: "book3",
  civilisational: "civilisational",
  art_soul_ai: "art_soul_ai",
} as const;

export type ExperimentChannel =
  | (typeof ExperimentChannel)[keyof typeof ExperimentChannel]
  | null;

export const ExperimentChannel = {
  chronic_pain: "chronic_pain",
  abi: "abi",
  dementia: "dementia",
  perinatal: "perinatal",
  eating_disorder: "eating_disorder",
  first_nations: "first_nations",
} as const;

export type OutputVoice = (typeof OutputVoice)[keyof typeof OutputVoice];

export const OutputVoice = {
  academic_only: "academic_only",
  ferrier_popular_only: "ferrier_popular_only",
  academic_first_then_ferrier: "academic_first_then_ferrier",
  raw_findings_only: "raw_findings_only",
} as const;

export type OutputFormat = (typeof OutputFormat)[keyof typeof OutputFormat];

export const OutputFormat = {
  report: "report",
  structured_data: "structured_data",
  annotated_results: "annotated_results",
  convergence_map: "convergence_map",
  frame_inventory_only: "frame_inventory_only",
  reflexivity_report: "reflexivity_report",
} as const;

export interface ProtectionsConfig {
  p1_falsification?: boolean;
  p2_eliza_output?: boolean;
  p3_meta_observation?: boolean;
  p4_independence_testing?: boolean;
}

export interface PositionPrivilegeBalance {
  state_admin?: number | null;
  credentialed_research?: number | null;
  community_curated?: number | null;
  indigenous_scholarship?: number | null;
  theoretical_tradition?: number | null;
  advocacy?: number | null;
  grey_practitioner?: number | null;
}

export interface Experiment {
  id: number;
  /** snake_case identifier (from artefact) */
  experimentId: string;
  project: ExperimentProject;
  status: ExperimentStatus;
  question: string;
  hypothesis?: string | null;
  expectedOutcomeTypes: string[];
  channel?: ExperimentChannel | null;
  patterns?: number[];
  protections?: ProtectionsConfig;
  evidenceTierThreshold: string;
  convergenceRequirement: string;
  includeLayers?: string[];
  framesExpected?: string[];
  framesExplicitlyExcluded?: string[];
  dissonanceBudget?: number | null;
  positionPrivilegeBalance?: PositionPrivilegeBalance | null;
  outputVoice: OutputVoice;
  outputFormat: OutputFormat;
  budgetCapAud: number;
  iterationCap?: number | null;
  timeCapSeconds?: number | null;
  requireHumanReview: boolean;
  observerNote: string;
  reflexivityQuestions?: string[];
  createdBy?: string | null;
  elapsedSeconds?: number | null;
  budgetConsumed?: number | null;
  currentIteration?: number | null;
  isTruncated: boolean;
  validationErrors?: string[];
  createdAt: string;
  updatedAt: string;
  completedAt?: string | null;
  artefactYaml?: string | null;
}

/**
 * JSON object if providing JSON directly
 */
export type ExperimentArtefactInputArtefactJson = { [key: string]: unknown };

export interface ExperimentArtefactInput {
  /** Raw YAML content of the experiment artefact */
  artefactYaml?: string;
  /** JSON object if providing JSON directly */
  artefactJson?: ExperimentArtefactInputArtefactJson;
}

export interface ExperimentUpdateInput {
  status?: ExperimentStatus;
  requireHumanReview?: boolean;
}

export interface ValidationError {
  message: string;
  errors: string[];
}

export interface ValidationResult {
  valid: boolean;
  errors: string[];
}

export type ExperimentFindingsConvergenceMap = {
  [key: string]: unknown;
} | null;

export type ExperimentFindingsPositionPrivilegeSummary = {
  [key: string]: unknown;
} | null;

export interface ExperimentFindings {
  experimentId: number;
  findingsMarkdown?: string | null;
  convergenceMap?: ExperimentFindingsConvergenceMap;
  positionPrivilegeSummary?: ExperimentFindingsPositionPrivilegeSummary;
  citations: string[];
  outcomeTypes: string[];
  generatedAt?: string | null;
}

export type ExperimentsSummaryByStatus = { [key: string]: number };

export type ExperimentsSummaryByProject = { [key: string]: number };

export type ExperimentsSummaryTopFramesItem = {
  frame?: string;
  count?: number;
};

export interface ExperimentsSummary {
  totalExperiments: number;
  byStatus: ExperimentsSummaryByStatus;
  byProject: ExperimentsSummaryByProject;
  totalBudgetConsumed: number;
  recentActivity: Experiment[];
  requireReview: Experiment[];
  topFrames: ExperimentsSummaryTopFramesItem[];
}

export type CrossExperimentViewConvergentFindingsItem = {
  theme?: string;
  experiments?: string[];
  strength?: string;
};

export type CrossExperimentViewDivergentFindingsItem = {
  theme?: string;
  experiments?: string[];
  tension?: string;
};

export type CrossExperimentViewFrameDistribution = { [key: string]: number };

export interface CrossExperimentView {
  convergentFindings: CrossExperimentViewConvergentFindingsItem[];
  divergentFindings: CrossExperimentViewDivergentFindingsItem[];
  frameDistribution: CrossExperimentViewFrameDistribution;
}

export type ReflexivityReportChannelBreakdown = { [key: string]: number };

export type ReflexivityReportProjectBreakdown = { [key: string]: number };

export interface ReflexivityReport {
  period: string;
  totalExperiments: number;
  dominantFrames: string[];
  underrepresentedPositions: string[];
  channelBreakdown: ReflexivityReportChannelBreakdown;
  projectBreakdown: ReflexivityReportProjectBreakdown;
  suggestedRebalanceExperiments: string[];
  positionPrivilegeAverage?: PositionPrivilegeBalance | null;
}

export type TemplateTemplateType =
  (typeof TemplateTemplateType)[keyof typeof TemplateTemplateType];

export const TemplateTemplateType = {
  channel_therapeutic: "channel_therapeutic",
  cross_cultural_validity: "cross_cultural_validity",
  frame_extinction_audit: "frame_extinction_audit",
  civilisational: "civilisational",
  meta_synthesis: "meta_synthesis",
  methodology_audit: "methodology_audit",
} as const;

export interface Template {
  id: string;
  name: string;
  description: string;
  templateType: TemplateTemplateType;
  artefactYaml: string;
}

export type ListExperimentsParams = {
  status?: ListExperimentsStatus;
  project?: string;
  channel?: string;
  search?: string;
};

export type ListExperimentsStatus =
  (typeof ListExperimentsStatus)[keyof typeof ListExperimentsStatus];

export const ListExperimentsStatus = {
  pending: "pending",
  running: "running",
  complete: "complete",
  failed: "failed",
  paused: "paused",
} as const;

export type GetCrossExperimentViewParams = {
  project?: string;
  channel?: string;
};