# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CRIA - Convergent Research Intelligence Architecture API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get a single template with prefilled artefact YAML
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Template with artefact YAML
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
components:
  schemas:
...
    Template:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        templateType:
          type: string
          enum: [channel_therapeutic, cross_cultural_validity, frame_extinction_audit, civilisational, meta_synthesis, methodology_audit]
        artefactYaml:
          type: string
      required:
        - id
        - name
        - description
        - templateType
        - artefactYaml